refactor(frontend): Content 필터 컴포넌트 중복 제거 및 공통화
두 다이얼로그 컴포넌트에서 동일한 필터 UI가 중복 구현되어 있었음 (content-details-dialog, content-group-details-dialog 각각 ~180줄씩 중복)

공통 컴포넌트로 분리하여 재사용성 개선:
- BooleanFilter: 더보기/귀속 포함 토글 통합
- ItemsFilter: 보상 아이템 멀티셀렉트 래퍼
- useWageFilter: 필터 상태 관리 훅 (items 변경 시 동기화 포함)
- ContentWageSection: 시급 정보 섹션 통합 컴포넌트
- RewardItem 타입 통합으로 중복 정의 제거

fix #272
