+++ src/frontend/src/domains/content/components/index.tsx
@@ -1,0 +1,1 @@
+export { ContentWageSection } from "./content-wage-section";
