+++ src/frontend/src/domains/content/filters/index.tsx
@@ -1,0 +1,2 @@
+export { BooleanFilter } from "./boolean-filter";
+export { ItemsFilter } from "./items-filter";
