+++ src/frontend/src/domains/content/hooks/index.tsx
@@ -1,0 +1,1 @@
+export { useWageFilter } from "./use-wage-filter";
