+++ src/frontend/src/domains/content/types.tsx
@@ -1,0 +1,4 @@
+export type RewardItem = {
+  id: number;
+  name: string;
+};
